install vue and tailwind
