+++ src/script/calculator.js
