+++ src/script/calculator.ts
