+++ src/shims-vue.d.ts
@@ -1,0 +1,6 @@
+declare module '*.vue' {
+    import { DefineComponent } from 'vue';
+    const component: DefineComponent<{}, {}, any>;
+    export default component;
+  }
+
